+++ src/main/kotlin/com/atlassian/performance/tools/aws/Ec2.kt
@@ -1,13 +1,15 @@
 package com.atlassian.performance.tools.aws
 
 import com.amazonaws.services.ec2.AmazonEC2
+import com.amazonaws.services.ec2.model.ImageState
 import com.amazonaws.services.ec2.model.InstanceStateName
+import com.atlassian.performance.tools.aws.ami.AmiImage
 import com.atlassian.performance.tools.aws.api.Resource
 import com.atlassian.performance.tools.aws.api.TerminationBatchingEc2
 import com.atlassian.performance.tools.aws.api.TerminationPollingEc2
 
 internal class Ec2(
-    private val ec2: AmazonEC2
+    private val ec2: AmazonEC2,
 ) {
     fun listExpiredInstances(): List<Resource> {
         val scrollingEc2 = TokenScrollingEc2(ec2)
@@ -26,4 +28,13 @@ internal class Ec2(
         }
         return instances.filter { it.isExpired() }
     }
+
+    fun listExpiredAmis(): List<Resource> {
+        return ec2.describeImages()
+            .images
+            .filter { ImageState.fromValue(it.state) != ImageState.Deregistered }
+            .map { AmiImage(image = it, ec2 = ec2) }
+            .filter { it.isExpired() }
+    }
+
 }